Refactoring and Server Code Fix Fixes #19, #11, #12 Moved API to api, marking first API version Switched to forge stable build
