+++ api/main/java/com/attributestudios/wolfarmor/api/IWolfArmorCapability.java
