+++ api/main/java/com/attributestudios/wolfarmor/api/item/IWolfArmorMaterial.java
