+++ api/main/java/com/attributestudios/wolfarmor/api/util/Definitions.java
