+++ src/main/resources/assets/wolfarmor/textures/models/armor/wolf_chainmail_1.png
